Move prepare_indicators to features module
